backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
langchain>=0.1.0
langchain-groq>=0.0.1
groq>=0.4.0
langchain-community>=0.0.20
langchain-core>=0.1.0
pypdf>=4.0.0
pdfplumber>=0.10.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
